+++ Travelers_Journal/CalendarTime.h
@@ -2,33 +2,33 @@
 #include <iostream>
 #include <fstream>
 
-const size_t NUMBER_OF_MONTHS = 12;
-const size_t MAX_NUMBER_LEN = 8;
+const unsigned NUMBER_OF_MONTHS = 12;
+const unsigned MAX_NUMBER_LEN = 8;
 
 size_t convertCharToInt(const char*);
 
 class CalendarTime
 {
-	size_t day;
-	size_t month;
-	size_t year;
+	unsigned day;
+	unsigned month;
+	unsigned year;
 
-	size_t maxDays[NUMBER_OF_MONTHS] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
+	unsigned maxDays[NUMBER_OF_MONTHS] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
 	bool isModified = true;
 
 	void validateDateByChangedCalendars();
 
 public:
 	CalendarTime();
-	CalendarTime(size_t, size_t, size_t);
+	CalendarTime(unsigned, unsigned, unsigned);
 
-	void setDay(size_t);
-	void setYear(size_t);
-	void setMonth(size_t);
+	void setDay(unsigned);
+	void setYear(unsigned);
+	void setMonth(unsigned);
 
-	size_t getDay() const;
-	size_t getMonth() const;
-	size_t getYear() const;
+	unsigned getDay() const;
+	unsigned getMonth() const;
+	unsigned getYear() const;
 
 	bool isLeapYear() const;
